move the logic from `doctor-general-cli` back up to `doctor-general`, now the distinction is much cleaner
