+++ packages/doctor-general/doctor-general.js
@@ -1,8 +1,9 @@
 export * from './comment'
+export * from './drgen'
 export * from './file'
 export * from './parse'
-export * from './renderer-markdown'
 export * from './renderer-jest'
+export * from './renderer-markdown'
 export * from './stats'
 export * from './string'
 export * from './text'
