+++ packages/doctor-general/log.js
@@ -1,5 +1,6 @@
 import { complextrace } from 'envtrace'
 export const log = complextrace('doctor-general', [
+  'core',
   'comment',
   'config',
   'file',
