+++ tools/doctor-general-cli/executable.js
@@ -1,5 +1,5 @@
-import { drgen } from './cli'
+import { cli } from './cli'
 import { fork } from 'fluture'
 
 // eslint-disable-next-line no-console
-fork(console.error)(console.log)(drgen(() => {}, process.argv))
+fork(console.error)(console.log)(cli(() => {}, process.argv))
